+++ great_expectations/rule_based_profiler/parameter_builder/simple_date_format_string_parameter_builder.py
@@ -16,10 +16,10 @@ from great_expectations.rule_based_profiler.types import (
     FULLY_QUALIFIED_PARAMETER_NAME_METADATA_KEY,
     FULLY_QUALIFIED_PARAMETER_NAME_VALUE_KEY,
     PARAMETER_KEY,
-    Attributes,
     Domain,
     ParameterContainer,
 )
+from great_expectations.types.attributes import Attributes
 
 logger = logging.getLogger(__name__)
