+++ great_expectations/rule_based_profiler/types/__init__.py
@@ -1,4 +1,3 @@
-from .attributes import Attributes  # isort:skip
 from .builder import Builder  # isort:skip
 from .domain import (  # isort:skip
     INFERRED_SEMANTIC_TYPE_KEY,
