+++ great_expectations/types/attributes.py
@@ -1,9 +1,12 @@
+import logging
+
 from great_expectations.core import IDDict
 from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.types import SerializableDotDict
 
+logger = logging.getLogger(__name__)
 
-# TODO: <Alex>If/when usage of this class gains traction, it can be moved to common general utilities location.</Alex>
+
 class Attributes(SerializableDotDict, IDDict):
     """
     This class generalizes dictionary in order to hold generic attributes with unique ID.
